reddit_app/api.py
from rest_framework import generics, permissions
from .models import Post, Vote
from .serializers import PostSerializer, VoteSerializer


class PostView(generics.ListCreateAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    permission_classes = [permissions.IsAuthenticatedOrReadOnly]

class PostDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    permission_classes = [permissions.IsAuthenticated,]

class VoteView(generics.ListCreateAPIView):
    permission_classes = [permissions.IsAuthenticated,]
    queryset = Vote.objects.all()
    serializer_class = VoteSerializer
    



